Core: Add secure and trusted flags to partition DB
Add a new flags field to the secure partition DB. Add two flags,
signalling whether the partition is secure and/or trusted. When a
partition ID is checked whether its partition is secure or not, the new
flag is used, instead of comparing the partition ID to the partition ID
of the non-secure partition. This creates the possibility for the TF-M
to be able to handle multiple non-secure partitions.

Change-Id: I0a99eeb1b439892a709b934de6d66f9598781a8c
Signed-off-by: Mate Toth-Pal <mate.toth-pal@arm.com>