SST: Set physical ID for logical block 0 on create
Changes sst_core_object_create so that, when an object is reserved in
logical block zero, the physical ID in the block metadata is updated to
the current scratch metadata block. This ensures the physical ID is
still correct once the metadata blocks are swapped at the end of the
function.

Change-Id: I6bc33191d1fc596963bf1c44b92fc5f710fb3b6b
Signed-off-by: Jamie Fox <jamie.fox@arm.com>