SST: Add reference read PSA SST API definition
This patch adds a reference read PSA SST API definition to read asset's
data on behalf of a given client ID from a secure partition. This
function is only accessible for secure partitions with specific
permissions in the asset policy table.
This API must not be accessible from the Non-secure Processing
Environment (NSPE).
To prevent the use of this API from the NSPE, this is not implemented
in the tfm_sst_api.c to trigger a linker error in case it is referenced
in the NSPE client code. In case, the NSPE client calls directly the
veneer, the SST service will detect it and return an error accordingly.

Change-Id: If033cbd2d24237b4b29cb8e034801ec2c3b84fc8
Signed-off-by: Marc Moreno <marc.morenoberengue@arm.com>