Build: Set proper response file flag for linkers
To avoid exceeding the command line length limitation, a linker supports
extracting parameters from a file instead of processing a long parameter
list in the command prompt. CMake puts linker parameters in a response
file (typically named as 'objects1.rsp') for linker if there are a lot
of parameters to be processed, then assign linker specified flags for
the linker to extract parameter from this file. CMake applies GNU 'ld'
option '@' as default flag if the flag is not set precisely for a
linker, which leads to potential linker error while using ARMCLANG:

Fatal error: L6002U: Could not open file
@CMakeFiles/tfm_s.dir/objects1.rsp: No such file or directory

Precisely set response file extracting flag for each linker to fix this
problem: '@' for 'GNUARM' and '--via=' for 'ARMCLANG'.

Change-Id: If2ce6644c16aafd0d8c7d8db01e42548ec81825e