Platform: Add new CMSIS Flash driver for Musca
- This patch replaces the former emulated and temporary CMSIS flash
driver (for Musca-A) with a new one which is based on the qspi_ip6514e
flash controller driver and library of the MT25QL flash memory device.
- Integrates the flash driver with MCUBoot and SST service
- The usage of the driver is limited when running from flash.
The erase functions and the flash command (STIG) based implementation
of the ARM_Flash_ProgramData() cannot be used and the initialization
has to be modified to use the default parameters instead of the
optimal.

Change-Id: Ida01cd327d197f0f1bd1adf913ea19aae9662ff3
Signed-off-by: David Vincze <david.vincze@arm.com>