Boot: Remove revert functionality support
This patch temporarily removes the revert functionality in case of
MCUBOOT_NO_SWAP and/or MCUBOOT_RAM_LOADING compiler switches.This
functionality would recover the previous working firmware version
in case of a faulty update. The reason is that the functionality is
not fully implemented and during the first boot (relative to flash
erase) the image trailer is deleted while the image will not be
confirmed by runtime, preventing the subsequent successful booting.

Change-Id: I45b2af44c49b4a51a749f5cf92aeb6004a5b28b6
Signed-off-by: David Vincze <david.vincze@arm.com>