Tools: Add asset policy table generation tool
This patch adds the SST asset policy table generation tool. The tool
generates the asset definition files, required for SST service to
build the policy table, based on the asset information and permissions
described in a YAML file.
This patch also updates the SST integration guide to explain how to
describe the asset information and permissions in the YAML file and
how to run the tool. It also changes the asset IDs to start from 1.

Change-Id: I64b3064a966821c825f457fe5110deb5e1f9ff7b
Signed-off-by: Marc Moreno <marc.morenoberengue@arm.com>