Doc: added documentation build support.
This change enables checking for availability of tools needed to build
the doxygen documentation (Reference Manual), and building the manual.

During the configuration phase CMake will search for doxygen, PlanUML,
dot, java and Latex tools.
Then it will "configure" the file doxygen/Doxyfile.in, and generate
the following targets:
  - doc_refman: to build HTML documentation
  - doc_refman_pdf: to build PDF documentation
  - install_doc: to install all build documentation files

If any mandatory tool is missing, CMake will issue a warning and not
generate the targets. Thus missing tools will not stop the user to build
the firmware, just remove the documentation build capability.
PDF generation is optional, if latex is missing, HTML documentation
can still be generated.

Change was tested on:
  -Win10 with native windows doxygen and MixTex for PDF generation.
  -Linux Mint 18.1 Serena (based on Ubuntu 16.04 - Xenial Xerus)

Change-Id: I25fb80bb8eee8e510c9bd55bffb5974d2360651f
Signed-off-by: Gyorgy Szing <Gyorgy.Szing@arm.com>