Platform: Repartition RAM area in Musca A1
This patch increases the secure RAM memory size (96KB) for the secure
environment and decreases it for the non-secure one (32KB) in order
to accommodate them to the memory requirements of regression tests in
Musca A1.
When it is not required to run the regression tests, the RAM memory
partition is the original one which is 64KB of RAM for each execution
environment.

Change-Id: I19336c875f740bb86cdf734d8f8779fa944f6710
Signed-off-by: Marc Moreno <marc.morenoberengue@arm.com>