NSPM: Update NSPM mechanism to get thread name
This patch updates the non-secure partition management (NSPM) to
use the public RTX osThreadGetName function to get the thread name
rather than use internal RTX functions.
It also gets the client ID from the thread name and passes it into the
SVC function to register the client ID to be assigned to the current
TZ context.

Change-Id: I269d2d09f21615e1fa9fc57d5c0239923a5f9fca
Signed-off-by: Marc Moreno <marc.morenoberengue@arm.com>