Platform: Fix incorrect AIRCR register config
This patch updates system_reset_cfg function to clear the AIRCR
VECTKEY value from reg_value before writing the new values in it.
That is required to guarantee that the 2 top bytes of reg_value
are 0x05FA, and the write is not ignored when the reg_value content
is assigned to AIRCR register.

Change-Id: Id13c169b7514368b72ce6206b28fd4a78d00b936
Signed-off-by: Marc Moreno <marc.morenoberengue@arm.com>