SST: Fix code logic in set_active_object_table
This patch fixies the code logic in sst_set_active_object_table when
table 0 is valid and table 1 is invalid. If both tables are valid an
extra evaluation is required to determine which table contains the
latest content based on NV counters, if SST_ROLLBACK_PROTECTION is
set, or swap counter.

Change-Id: I3a751c003b5c7b272005ecacd26f607f37978e1e
Signed-off-by: Marc Moreno <marc.morenoberengue@arm.com>