Platform: Update headers in system_cmsdk_musca.c/h
This patch corrects the copyright year in the system_cmsdk_musca.c/h
files for Musca-A boards.
Adds the github link and sha hash to the header from which the files
are derived.
Incorrect copyright year in system_cmsdk_musca.h/c files for musca_a
was introduced in previous commit (sha 788036c2b)

Change-Id: Ic7e5adff5f8ccba05584bfa5272fbf703e964bd4
Signed-off-by: Avinash Mehta <avinash.mehta@arm.com>