Core: Fix thread list handling
The function update_running_head(...) updates the pointer pointing to
the highest priority runnable thread.

The old implementation updates RUNN_HEAD pointer to the new running
thread, if the new running thread has the same or higher priority then
current RUNN_HEAD. However this is not the correct behaviour.

If a thread with the same priority as a running thread becomes running,
then RUNN_HEAD is updated to this thread. Then if this second thread
gets blocked, then the previous thread will never be scheduled, as the
RUNN_HEAD pointer is past it, and the search always starts at
RUNN_HEAD.

Fix: Change tfm_thrd_set_status to update the RUNN_HEAD to the new
running thread, if it has higher priority than the thread at RUNN_HEAD.
In other cases start the search from the beginning of the thread list.

Change-Id: I3738fa2e728d731b5b2553ad64ba232b47509c5a
Signed-off-by: Mate Toth-Pal <mate.toth-pal@arm.com>