Doc: update supported tool versions
This document update expands and adds more details about the supported
build environments and supported version of build tools.
There was extensive test executed to validate the build system with the
newly added tool versions. While the intention is to keep the range of
tool versions as limited as possible, default tool versions of widely
used build environments made the list somewhat extended.

Change-Id: I8711fe211c20e39126f953e5518ed2a21ab9eb7f
Signed-off-by: Gyorgy Szing <Gyorgy.Szing@arm.com>