Platform: Extend NV counter interface
This patch adds the tfm_plat_set_nv_counter() function to the NV counter
interface for supported platforms: AN521, AN519, Musca-A and Musca-B1.
To set a NV counter to a specified value can take a lot of time and
leads to early flash wear out when incrementing the counter by one.

Change-Id: I8702de54d373aff3201669540c541a24815cf09b
Signed-off-by: David Vincze <david.vincze@arm.com>