Boot: Use overwrite-only mode for image upgrade
- The swapping image upgrade is a two-step process (by default). First
it performs a "test" upgrade and boots the new image. Then the image can
mark itself as "OK" at runtime by setting the image_ok flag in the
flash. When this happens, the swap is made "permanent" and mcuboot will
then still choose to run it during the next boot. Currently TF-M does
not set the image_ok flag, therefore the bootloader will always perform
a "revert" (swap the images back) during the next boot. For this reason,
this patch sets the overwrite-only mode as the default image upgrade
strategy. However, this strategy does not provide the opportunity to
restore the previous image if the device crashes with the new firmware.
- Update secure boot documentation.
- Fix format specifier error in overwrite-only code.
- Fix the bootloader output message by reorganizing the
preprocessor directives in case of non-swapping firmware upgrade.

Change-Id: Ida1c7c7fc198e393e79d178269826cbfa052395d
Signed-off-by: David Vincze <david.vincze@arm.com>