Attest: Fix the detection of option flags
Option flags for test purpose can be passed in to
attestation service. It is encoded in a special challenge
value. In this case challenge object must be 64 bytes
long and the last 60 bytes (4-63) must be 0 value then
the first four byte (0-3) handled as the option flags.
Detection of these flags was mishandled if challenge
object is not 64 bytes long.

Change-Id: I457ab8d52236c2261ca12b0ae0f341c4c15976b2
Signed-off-by: Tamas Ban <tamas.ban@arm.com>