Boot: Save image sequence number to image trailer
Overload the swap_type field in image trailer to store as an addition
the image sequence number. It indicates which image's swap was
interrupted. It is required by multi image boot to determine which
image the trailer belongs to if boot status is found on scratch area
when the swap operation is resumed.

Change-Id: I1a695fa7b51ccf7fb7969496517617d76bc44c02
Signed-off-by: Tamas Ban <tamas.ban@arm.com>
Co-authored-by: David Vincze <david.vincze@arm.com>