Attest: Increase token size macro
In case of multiple image boot (secure and non-secure
image is signed independently) the attestation token
will be bigger because the SW_COMPONENT array contains
two elements (S and NS) instead of one (single combined
S+NS). The buffer size must be increased to be able to
accommodate this bigger token.

Change-Id: I3b53c6fbb1418c608a7a8039f6e2f9aadfae619e
Signed-off-by: Tamas Ban <tamas.ban@arm.com>