Attest: Refine test suite to support multi-image boot
The expected values of claims must be modified in
attestation test suite to support the multiple image
boot scenario. In some cases values were just dummy data.
Most cases just the presence of the claim is tested and
its value is not. Because the value can dynamically change
per build (image hash value).

Change-Id: I1faf06c3c54f0c07abad5fa91771d7745b5d15f3
Signed-off-by: Tamas Ban <tamas.ban@arm.com>