Boot: Add hardware key support to MCUBoot
PSA Trusted Boot and Firmware Update specification requires
the support of at least one immutable root of trust public key
(ROTPK) for firmware verification. This key is provisioned to
the SoC during manufacturing. This change enables MCUBoot to
get the ROTPK hash from the SoC and compare with calculated
hash of public key from image manifest to validate the key
which is used for firmware authentication before usage.

Change-Id: I68cf4104bc2984fa8ad4c902d2a9a36064939f1e
Signed-off-by: Tamas Ban <tamas.ban@arm.com>