Core: Tweak 'update_running_head()' behaviour
This function could be called in SVCall functions, it searches for the
next running thread if the 'Running Head' thread status has been set to
non-running. Performing a searching in SVCall is unnecessary and costs
more expensive SVCall execution time. Reset the searching start head to
thread list and return in this case, let scheduler in PendSV handles the
searching.

Change-Id: Idc795496f1fcb03e139e2d37d9ab46bf52c9b1e1
Signed-off-by: Summer Qin <summer.qin@arm.com>