Boot: Enable multi-image boot
This patch adds the capability to handle multiple firmware images,
to update them independently. It separates the completion of aborted
image swap operations and the update of images even more, as these
should be happening at different stages of the boot process according
to the design proposal of the multiple image support:
https://github.com/JuulLabs-OSS/mcuboot/pull/317.

Also update the code related to rollback protection and the saving of
the SW component's boot status info.

Change-Id: I0f2159f40b47077fe22391ffca0f0376218d1cb1
Signed-off-by: David Vincze <david.vincze@arm.com>