Interface: Refine OS wrapper APIs
For the following OS wrapper APIs
    os_wrapper_thread_new
    os_wrapper_thread_get_id
    os_wrapper_thread_get_priority
    os_wrapper_semaphore_create
    os_wrapper_mutex_create
They return the corresponding IDs or priorities on success and
OS_WRAPPER_ERROR on failure.
This makes the "OS_WRAPPER_ERROR" value invalid for thread IDs,
mutex IDs...etc on all OSes.
But it could be a valid value on some OSes.

This patch fixes this problem by
1. Changing mutex, semaphore and thread IDs to handles
2. For os_wrapper_thread_get_priority, putting outputs to
   parameters and returning only success or error.
   For the others, treating NULL return as error and others for
   valid handles.

Change-Id: I19aa0415ce0ba4388fb96602fea8f10f702d8c45
Signed-off-by: Kevin Peng <kevin.peng@arm.com>