Tools: Add output argument for manifest parsing tool
This patch allows user to specify a custom location to put the files
generated by the manifest parsing tool.

Note that the argument is optional. The tool can be used as before
without argument and the functionality is not changed.
For detailed usage, try "python3 tfm_parse_manifest_list.py -h"

Change-Id: I26e77314c76ac68787c30a135cefeb414f9a941a
Signed-off-by: Kevin Peng <kevin.peng@arm.com>
Co-authored-by: Shawn Shan <shawn.shan@arm.com>
Co-authored-by: Alan DeMars <ademars@ti.com>