Build: Fix NS image linking error for Armclang
This patch provides a workaround for a certain linker error which was
first observed on Musca-A and Musca-B1 platforms when building the
Non-secure image with Arm compiler version 6.11 or 6.12. The patch
simply replaces the .ANY module selectors with the * module selector
in each scatter file. This issue was introduced with 6.11,
more information: https://www.keil.com/support/docs/4150.htm.

Change-Id: Ia862205604172c2e92cf69074e8cc0834787e696
Signed-off-by: David Vincze <david.vincze@arm.com>