Docs: Fix environment setup instructions
The example in the documentation for setting the PLANTUML_JAR_PATH
environment variable is wrong and makes cmake unable to find the tool.
As a fix, the quotation is removed.

The problem is the example has the value to be set quoted while the
path is ~ relative. As a result, the shell will not resolve
the ~ character to the home path, and the environment variable
will contain the path in the verbatim ~ relative form.

When the build-system uses the value, it gets passed to the
kernel as is, but the kernel can not resolve ~ as ~ referring
to the home directory is a shell feature.

Signed-off-by: Vikas Katariya <Vikas.Katariya@arm.com>
Change-Id: I6d4807b5855c4bb74fe9c58f99e996dea202c4e2