Build: Fix pdf document build
CMake fails to generate proper rules for PDF generation both for the
Reference Manual (doxygen) and for the user Guide. The problem is
caused by CMAKE_MAKE_PROGRAM not being set when the document creation
targets are defined. As a result CMake generates "empty" targets which
do not execute the final step of PDF generation. The problem was
triggered by project re-organization.

To fix the problem document generation files are included after the
project() command. Also the defined state of CMAKE_MAKE_PROGRAM is
verified and a meaningful error message is printed to make the build
more robust.

In addition
- I made some minor changes to make the document build related cmake
  files more consistent.
- Fixed spelling issues.

Change-Id: Id5d8dc756a50b3fee8966b51461ad3607d2027b4
Signed-off-by: Gyorgy Szing <Gyorgy.Szing@arm.com>