SST: Remove object system lock and ready variable
Removes the sst_global_lock and sst_global_unlock calls from the SST
object system, along with the sst_system_ready variable, as they are no
longer required. Ensuring a single thread of execution in a service is
the responsibility of the SPM, while ensuring that the service is
initialised is now handled by the SST request manager, separately for
IPC and library models.

Change-Id: I8d51e0ff4198aa76e026e4e63cfd24c468b88740
Signed-off-by: Jamie Fox <jamie.fox@arm.com>