Attest: Use boot record TLV in IAT token creation
Use the build time generated measured boot record structure which is
already CBOR encoded for the creation of the SW_COMPONENTS claim, but
also keep the backward compatibility.
Set the default value of ATTEST_BOOT_INTERFACE to 'CBOR_ENCODED_CLAIMS'
so the bootloader will copy (after a successful image authentication)
the content of the boot record TLV to the shared data area instead of
the individual SW component claims.

Change-Id: I33f1a89f8b2befed914c6ea9c77b1e0f896d27d7
Signed-off-by: David Vincze <david.vincze@arm.com>