Attest: Verify access policy to shared data area
In shared data area the data items are mixed and addressee can
be various secure partitions or even SPM. Introduce an access
policy check to limit which partition has access to which data
items in the shared data area. This check mitigates the risk
to disclose sensitive data to unauthorized secure partitions.

Change-Id: Ibab2ff46e091c3786565be69c94cd9f02c664f3a
Signed-off-by: Tamas Ban <tamas.ban@arm.com>