Boot: Provide signer ID even when using HW keys
The signer ID claim was not copied to the shared data area when HW keys
were used for image verification (that was because the key-hash TLV was
not present in the image manifest).
This patch provides a fix to write the signer ID to the shared data area
when MCUBOOT_HW_KEY is set to True by retrieving the full public key
from the image manifest and calculating its hash on-the-fly.

Change-Id: I9e4b968a40ecaa21fd18b536525a569ad6539fd3
Signed-off-by: David Vincze <david.vincze@arm.com>