Support CMSIS_device_header macro provided in CMSIS
This patch creates the tfm_hal_device_header.h to make the device header
in TF-M compliant with CMSIS.
CMSIS uses a preprocessor define CMSIS_device_header that reflects the
actual device name.
With this patch, if the CMSIS_device_header is defined, then the device
header defined by the macro will be included, otherwise the original
cmsis.h.
To be noted that the existing platform ports are not affected.

Change-Id: If24337be37188335ea514769fd90de75365925d9
Signed-off-by: Kevin Peng <kevin.peng@arm.com>