ITS: Add support for multiple filesystem contexts
Refactors the ITS flash filesystem to support multiple contexts, to
permit multiple users of the filesystem. Removes the internal static
context from the filesystem and replaces it with a context to be
allocated by the user and passed as an argument.

Also updates the flash layer to support multiple flash devices, with
the ability to associate a flash device with a filesystem context.

This will enable Protected Storage and Internal Trusted storage to
share the same filesystem code.

Change-Id: I7f2f2a6520816f36cdf0cd1e4b87a24c7b6ff5f9
Co-authored-by: Tudor Cretu <tudor.cretu@arm.com>
Signed-off-by: Jamie Fox <jamie.fox@arm.com>