Build: set SST_TEST_NV_COUNTERS for SST regression
The SST_TEST_NV_COUNTERS should be set to ON only when SST regression
is ON for all level of isolation.
This patch first moves out the test configurations out of
test/CMakeLists.txt to a seperate file for CommonConfig.cmake to
include and then set SST_TEST_NV_COUNTERS to on when SST regression
is on.

Change-Id: Icc7c7086023d73f2570ef27233abc2cdc0c480e0
Signed-off-by: Kevin Peng <kevin.peng@arm.com>