ITS: Refactor filesystem write function
Refactors the filesystem write function to support writing data from
an offset in the existing data, instead of having to always overwrite
all existing data.

Simplifies the dblock_write_file and dblock_cp_remaining_data functions
into a single function, as it was mandatory to call them in sequence
anyway.

Change-Id: I5162276b4e528a2235d0ceb73272538db5f98f8a
Signed-off-by: Jamie Fox <jamie.fox@arm.com>