Tools: Add arguments for manifest parsing tool
This patch adds two more optional arguments to the mainifest parsing
tool:
1. The manifest list file
    The custom secure partition manifest. It can replace or be appended
    to the default one.
2. The generated file list
    The custom generate file list. It can replace or be appended to
    the default one.

Note that these arguments are also optional individually and the default
values without arguments are not changed. The tool can be used as before
and the functionality is not changed.
For detailed usage, try "python3 tfm_parse_manifest_list.py -h"

Change-Id: I9d4340a2e1ac014090c35b3b55a330105ec8b7e9
Signed-off-by: Kevin Peng <kevin.peng@arm.com>
Co-authored-by: Alan DeMars <ademars@ti.com>