Core: Trigger a system reset for programmer error
According to PSA FF 1.0.0:
"The SPM cannot directly terminate the execution of an individual NSPE
task when a PROGRAMMER ERROR is detected because NSPE tasks are
typically managed by NSPE firmware.
Arm recommends the specified errors codes are also sent to any NSPE
management firmware. The NSPE management firmware can then decide to
pass those error codes back to the calling task, or to use its own
functionality for terminating an execution context.
For simple systems, it can be more appropriate for the SPM to restart
the entire system when a programmer error is detected in either the SPE
or NSPE."
The SPM will restart the entire system when a programmer error is
detected in either the SPE or NSPE in the first stage.
So in the first stage, SPM will restart the entire system when a
programmer error is detected in either the SPE or NSPE. It will be
enhanced in the future.
Note, tfm_core_panic() is the only function to handle PROGRAMMER ERROR.

Change-Id: I52a5e9af6dec1dbd8ec150a370d7bad1a0bcd36b
Signed-off-by: Edison Ai <edison.ai@arm.com>