Boot: Enable overwriting upgrade with 0 size scratch
Previously the bootloader always checked whether the size of the scratch
area is sufficient. In case of the overwrite upgrade strategy this check
is not necessary since it doesn't use the scratch. However, if the
FLASH_AREA_SCRATCH_SIZE was set to 0 it caused an error which prevented
any image upgrade attempts. This issue has already been fixed in the
upstream MCUBoot repository (commit hash:
74aef312df0d8f426a8ac04f359a377f7674dc7c). This patch simply copies the
fix from there and applies it to the forked MCUBoot in TF-M.

Change-Id: I7ffb8356239a1cf882be9c9c3a4e148b15bc7fa2
Signed-off-by: David Vincze <david.vincze@arm.com>