Crypto: Refine how crypto modules are selected
Crypto modules can be only enable/disable by setting the
CRYPTO_XXX_MODULE_DISABLED to 0/1.
This patch refines how the modules are selected. With this patch,
the crypto modules can be enabled/disabled by setting the
CRYPTO_XXX_MODULE_DISABLED to 0/1, OFF/ON or FALSE/TRUE, not
limited to 0/1 anymore.

Change-Id: I4f5ea82805760883639950cd6031cbf658cb0bfc
Signed-off-by: Kevin Peng <kevin.peng@arm.com>