Attest: Accept empty shared data area
Introduce the BOOT_DATA_AVAILABLE CMake variable to indicate whether
boot data is available in the shared data area (between the boot loader
and runtime firmware). If it's false the content of the shared data area
will be ignored and thus all the tests will pass for example when TF-M
is used with a boot loader that doesn't provide any boot data or when a
boot loader is not used at all.

Change-Id: Ic6d32cfdc8741018c6668692d64f81236006b593
Signed-off-by: David Vincze <david.vincze@arm.com>