Platform: Add support for cy8ckit-064b0s2-4343w board
Target cy8ckit-064b0s2-4343w is the actual psoc64 target.
The board has to be provisioned with Cypress Bootloader 1.1.0.424
(comes with cysecuretools 1.3.3) or newer.

Removed obsolete policy files and added a new policy for the board
that can be used for provisioning and signing the images:
policy_multi_img_CM0p_CM4_debug_2M.json

Signed-off-by: Andrei Narkevitch <ainh@cypress.com>
Change-Id: I8074d07ec6d589e6f893eea139da2809786a5403