Crypto: Remove const qualifier from secure API calls.
This patch removes the const qualifier from API method
calls of the CRYPTO api.

Depending on toolchain, local const variables can be
in RO data or on the stack. The TF-M linker script places
the default RO section in privileged memory.

This means that the caller if it is in  unprivileged mode,
cannot access this data for making the call to crypto service.

This issue was exposed when SST was moved to APRoT as part of
PSA L2 isolation. With this patch, the local variables are
placed on the corresponding stack, which is managed
by the Partition.

Change-Id: Ie0bb15ec0c0cbaa91178b9820484ac6c25d959fe
Signed-off-by: Minos Galanakis <minos.galanakis@arm.com>