Platform: Warm reset when NMI happens on musca_b1
It has been observed that secure watchdog is not able
to reset the system on the second timeout on musca_b1
board. Hence performing the warm reset as part of
watchdog interrupt (Take place in NMI). This change
will help the PSA FF test to recover from the panic
and will be able to jump to next test.

Signed-off-by: Jaykumar Pitambarbhai Patel <jaykumar.pitambarbhaipatel@arm.com>
Change-Id: I30436fcb066c681a6d7e8f20d8da22d9c9aee88c