Platform PSoc64: Flag PC=4 as secure
The Cypress bootloader allocates the Protection Contexts
to the various modes, and it specifies that PC=4 is
available as a secure context (it configures the
PROT_SMPU_MS0_CTLand PROT_SMPU_MS14_CTL registers
accordingly).
Even though we don't currently use PC=4 in TFM, the code
should match the system configuration.

Signed-off-by: Chris Brand <chris.brand@cypress.com>
Change-Id: Ic84548d8ce15bf5474a5d75b6fc38bd43f81029d