Doc: Update implication of using flash when running regressions tests
This patch updates the implication of using flash when running
ITS and SST regression test. The tests create WRITE_ONCE test assets
during the test execution. This means that when the test runs again
the test asset creation fails as the WRITE_ONCE assts cannot be created
again.

Also, if there is latent data in the flash which is not compatible with
the SST FS data, the tests will work correctly. This means that the flash
needs to be erased before running the tests.

Signed-off-by: Soby Mathew <soby.mathew@arm.com>
Change-Id: I201bd6299cb59fdf53fac3d68025c41966283605